Refactor out the player health UI logic from the player class
