+++ Assets/ScriptableObjects/Player/MaxPlayerHP.asset.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 31a3941ac5a943f40aa03d3cbf4e5bb3
+guid: 8848d63fa4a86d24eb0593f68351ad87
 NativeFormatImporter:
   externalObjects: {}
   mainObjectFileID: 11400000
