+++ Assets/_Scripts/Player/Player.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 36d082fcf3333fc41b1f6daa9c42faac
+guid: 626df3021e5470342b781384255116a5
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
